Moving cleanup logic to probr.GetAllProbeResults
